user_delete 구현

+++ src/main/webapp/WEB_INF/views/user_info.jsp
@@ -14,10 +14,12 @@
 			<sec:authorize access="isAuthenticated()">
 				<sec:authentication property="principal" var="principal"/>
 				<h2>${principal}</h2>
+				<a href="/user/before_edit?username=${principal.username}">회원정보 수정</a>
+				<a href="/user/delete?username=${principal.username}" onclick="return confirm('정말 삭제하시겠습니까?')">회원 탈퇴</a>
 			</sec:authorize>
-			<a href="/">돌아가기</a>
-			<a href="/user/before_edit?username=${principal.username}">회원정보 수정</a>
-			<a href="/user/delete?username=${principal.username}" >회원 탈퇴</a>
+			<p>
+				<a href="/">돌아가기</a>
+			</p>
 		</div>
 </body>
 </html>
